build: configure bin name in spec file
This is effectively a no-op when building qpc from this upstream repo.

This change exists only because I made a similar change to the downstream
discovery-cli fork where the `binname` value *is* different, and I wanted
to apply the same general change here to minimize the overall differences
between our upstream and downstream spec files.