Document that qpc source edit --port 0 inadvertently fails
`qpc source edit --name exists --port 0` will always fail because
`args.port` is used in boolean context and 0 is False. Add a test case
that would pass if the issue was fixed, and mark it as skipped with
explanation what is wrong.

See also https://github.com/quipucords/camayoc/pull/583